BACI X-mas present: deleted crack propagation
This was currently unused and unmaintained.
The implementation was spread across BACI and
would have required a lot of work in maintaining
and transition to the new structural time integration.

This is the first topic to be systematically and
deliberately removed (but SVN obviously never forgets).
To handle those cases, we have a new wiki-page
"BACI - phased out code" containing outdated topics,
when and why they were removed etc. If you still find
crack-related functionality in BACI, you can delete it
but please add a brief comment to the wiki-page, that
this commit was related to crack propagation.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22725 060e22d8-9d1c-0410-868a-e3a950d63d95