Gen-Inverse Analysis ====================
If inverse analysis finds not a new descent direction, so now longer converges,
up to now calculation continued till maximal number of iteration was reached. 
But nothing new is calculated.
Now inverse analysis stops, if no new descent direction is found.


Changed output of fitted parameters:
Each output-line prints the (grad)error with the parameters found for the next 
iteration. In case this parameters are not negative, they are used in the next iteration.
To check that parameters got negative, this mismatch stays.

But added a new output-line at the end, which shows the (current) lowest error with
the according parameters. So at the end of the iteration, the last line shows the
final parameters with the right error.
Special cases:
- grad error reached minimum: nearly nothing changed (parameters stay the same at the end)
- maximum iteration reached: handle final parameters with care (the same as before)
- inverse analysis finds no descent direction: now final parameters are parameters with the lower error (up to now, last line contained parameters, with higher error than second last line)
- negative parameters found: final parameters are last valid parameters


Deleted unused variables.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22736 060e22d8-9d1c-0410-868a-e3a950d63d95