Deleted inv_analysis (lung inverse analysis) ============================================
This was Sophie's "lung inverse analysis".
This type of analysis was not used for many years.

Reason for removing:
The features of this inverse analysis are also in gen_inv_analysis (just more general).
So a lot of redundant code existed.

If you need it: 
Find one of Sophie's dat-files and get it from the repository.

Documented also in:
https://lnm.wiki.tum.de/BACI+-+phased+out+code

Deleted also input-parameters, which were just necessary in lung-inv-analysis.
Deleted according testcase.

Approved by Lena & Sebastian Kehl.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22737 060e22d8-9d1c-0410-868a-e3a950d63d95