enabled most of our elements and materials in new structural time integration
Introduced a small converter function, that writes info from
the DRT::ELEMENTS::ParamsInterface back to the Teuchos::ParameterList.
This is not the way to keep it, but it is a clean way to have
both time integration strategies working at the same time
since the interface between elements and and time integrator is
still the same. 

Once we deleted the old time integration, this conversion will be
removed and all data at element level will be read through the
DRT::ELEMENTS::ParamsInterface to avoid ParameterList look-ups.

moved some test cases to new time integration


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22739 060e22d8-9d1c-0410-868a-e3a950d63d95