A new folder for documenting phased-out code ============================================
In the directory doc_removed_code we have a file major_changes.txt were we
list all major features of BACI that get phased out. Whenever you delete a
larger module, please list the revision of the change, the date and a summary
of why the code was deleted. This list is used as a documentation of features
we do not maintain actively any more but might or might not contain code that
can get re-activated in case a new project demands for it.

Please put new developments in the file major_changes.txt at the top in
reverse chronological orders, newer changes come in at the top after the
general description line.

Keep the description of a change in major_changes.txt to a few lines. If you
need more extensive documentation, please add a file in
doc_removed_code/details/2017_SummaryNameOfChange.txt where you list the
details. Please mention this file in major_changes.txt

There is also a file todo_list.txt where we collect topics to be done in the
near future but where we currently do not have resources for. These topics
could be part of the next cleaning week, student projects, etc.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22936 060e22d8-9d1c-0410-868a-e3a950d63d95