Switch BACI C++ standard to C++11 =================================
This commit makes the C++11 standard the default in BACI. This was previously
only active with the '--experimental' configuration. Thus, you are now
officially allowed to use features of C++11, such as lambdas, auto
declarations, significant extensions to the standard template library, and
much more. See e.g. here for an overview:
https://en.wikipedia.org/wiki/C%2B%2B11

The old configuration files for FC16 and kaiser that do not support C++11 have
been deleted with this commit. If you still need them, extract them from the
svn history, but beware that it is your responsibility to port code that is
not compatible with pre-C++11.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23048 060e22d8-9d1c-0410-868a-e3a950d63d95