In extension of r23107, allow largers matrices and vectors in all assemble functions than the size of the indices to be assembled.
The check whether the size of the location vector is equal to the length of the element 
matrix is replaced by the check for whether the size of the location vector is smaller 
or equal the size of the element matrix. This allows us to have elements which do not 
assemble their stiffness contribution to certain dofs in particular derived classes.

Also do the check both in DEBUG and RELEASE mode because it is very cheap and
fundamental to the functionality.

Patch partly prepared by Andreas Rauch.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23137 060e22d8-9d1c-0410-868a-e3a950d63d95