correction with respect to recent revisons 23110 and 23111
The failure of the low-Mach-number test cases was due to an incorrect computation of body forces for the residual of the fluid momentum equation when computing the subgrid-scale part of the velocity in the scalar transport solver. Hopefully, this is now fixed, and the cases are will be tested positively again. Due to the rather indirect impact, it took some time to find that out eventually ;-). Fortunately, however, those test cases featured both body forces AND the inclusion of the subgrid-scale velocity; otherwise, it would have gone unnoticed for the time being most probably.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23142 060e22d8-9d1c-0410-868a-e3a950d63d95