* linkage of beam elements - triad initialization *
Changed the way how triads of newly created linking elements are
initialized. The second base vector now is determined from the cross-
product of the first base vector and either the global x- or
y-direction, ensuring that the two vectors are never coaxial and thus
the cross-product yields a feasible second base vector.

Before, the Smallest Rotation procedure was applied which would cause
problems in the unlikely case of antiparallel vectors.

no change in results of tests

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23158 060e22d8-9d1c-0410-868a-e3a950d63d95