changes in test 'particle_27_gravity_MD':
- changed PROBLEM TYP to Particle
- set MOVING_WALLS to No
- changed results

explanation:

So far in this test the displacements of structural nodes were solely constraint in x and y direction,
displacements in z direction remained unconstrained resulting in displacement of order 1e-15.
Now walls are not moving, hence all structural displacements are zero leading to changes in the results
due to a large timestep in the explicit euler particle time integration.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23169 060e22d8-9d1c-0410-868a-e3a950d63d95