Cleanup of variable expression functions. Input is now analogous to the standard FUNCTs and e.g. looks like
COMPONENT 0 VARFUNCTION S0-2/pi*atan(0.01*dp) NUMCONSTANTS 1 CONSTANTS S0 0.3

The VARFUNCTION allows to define and set constants (in the example: S0) and to utilize variables which are initialized, set and used by external algorithms (in the example: dp, which is used by some poro materials). Component 0 is optional if a scalar functions is defined.

Patch by Moritz Thon.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23175 060e22d8-9d1c-0410-868a-e3a950d63d95