Patch by Sudhakar
files related to quadrature compression (QC), see comments by Sudhakar below.

---------------

The direct divergence method gives accurate integration rule, but may yield sometimes more than 1000 points. The advantage of QC is that all the volume-cells will have 56 integration points, irrespective of their shape. QC involves two steps:

1. LU decomposition of the Vandermondte matrix -- to get the subset of the original quadrature obtained from direct divergence

2. Solve a system of equations to obtain the modified weights.

The preprint of the attached paper explains the algorithm and also the efficiency gain (which is not very high, but still can be useful).

As far as the implementation, I define a flag "QUADCOMP" in cut_tolerance.cpp, and if this flag is on, we perform compression. Ideally, we should have an input parameter, but since I did this in the last moment I did not do it. Sorry about this!

Moreover, I am currently performing compression if the no. of integration points in the base quadrature is more than 56. This may not be ideal, because let us say if the no. of points is 60, performing all the operations to eliminate 4 points may not be of worth doing. So one can also raise the minimal no. of points to say 75.

---------------


By default, this technique is switched off. It needs further validation, as the standard integration and quadrature compression yield slightly different results.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23272 060e22d8-9d1c-0410-868a-e3a950d63d95