Write output for each Newton step during one load step into an extra output file.
This class is designed for debugging purposes only. The corresponding
input parameters are summarized in the section "---IO/EVERY ITERATION".
The implementation is meant to work as plug-in solution for all kinds of
problems. Anyway, the first implemented use can be found in the structural
time integration, such that it can be seen as a first working example.

In general, if you consider to use this class, you will have to construct
an object and call Init() and Setup() in consecutive order. Pass the
output-writer of your problem to the Init() function. This output
writer will be copied, such that the every iteration output should generate
the same output each Newton step just as the parent writer does in each
time/load step.

Secondly, you have to give this class access to your Output routines.
All methods which have to be accessed are summarized in the
EveryIterationWriterInterface class. Just add the abstract interface class
as base to your inheritance hierarchy of the class which contains these
functions. See STR::TIMINT::Base for an example.

Finally, such that the output is actually written you have to call the
public class methods at the correct point in your implementation.
The method InitNewtonIteration() should be called before you enter your
Newton loop. AddNewtonIteration is supposed to be called just in the end
of each Newton iteration.

You can find your every iteration output in an extra directory located in
your specified output folder. For example: If your output path is
"../o/fancy_sim", the every iteration output is written to
"../o/fancy_sim_every_iter/fancy_sim_step_<STEP_NP_NUMBER>". Each
step gets an extra control, result and mesh file. It is also possible to
restrict the output to one single time/load step. If you perform different
runs of the same simulation and the same load/time step, this is also
supported. Just use the input parameter RUN_NUMBER. If a positive number
is provided the result of each run can be found in the directory
"../o/fancy_sim_every_iter/run_<RUN_NUMBER>/".

- New input parameters ---IO/EVERY ITERATION

  OUTPUT_EVERY_ITER         yes/no   : Switch the every iteration output on/off.
  RUN_NUMBER                <int>    : Create a new folder for different runs of
                                       the same simulation. If equal -1, no folder
                                       is created.
  STEP_NP_NUMBER            <int>    : Give the number of the step (i.e. step_{n+1})
                                       for which you want to write the debug output.
                                       If a negative step number is provided, all
                                       steps will be written.
  WRITE_OWNER_EACH_NEWTON_ITER yes/no: If yes, the ownership of elements and nodes are
                                       written each Newton step, instead of only once
                                       per time/load step.

- Minor changes in the IO::OutputControl and IO::DiscretizationWriter.
  (1) Added copy constructors.
  (2) Generalized NewResultFile method.

- Fixed a clang compliler warning in the NOX framework.

- Modified a existing test case to test the new output functionality.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23292 060e22d8-9d1c-0410-868a-e3a950d63d95