added a warning in WriteAccessDispNp() in the new structural time integration: ------------------------------------------------------------------------------ Even though this method is called WriteAccess you can NOT modify the displacement state that is evaluated. This is due to the fact, that when calling Evaluate() on the structure field, the NOX group overrides the current displacement state with the solution vector sored inside the group.
In many (most?) cases WriteAccessDispNp() is simply used to get a pointer to
a non-const Epetra_Vector, while actually a const would be acceptable.
So keep in mind to put const statements whenever you are sure that you only
want read access which would allow to use the DispNp().

Of our current test cases only some immersed_* seem to use this function.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23327 060e22d8-9d1c-0410-868a-e3a950d63d95