Minor modification of some function calls. Basically, missing (const) references have been added, since many input parameters were passed by value. Now const references are used, instead. Speed-up of the element evaluation by about 5-7 percent.
Note that values of type bool, char, int, unsigned, etc. (smaller than 64-bit)
can and should be passed as copies, if they are not supposed to be changed. On
the other hand, larger objects should be passed by (const) reference (size of a
double), otherwise an additional constructor call and storage allocation slows
down your simulation.

No changes in functionality.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23339 060e22d8-9d1c-0410-868a-e3a950d63d95