Fix parallel distribution of elements on restart for non-standard ghosting
Patch by Andreas Rauch.

When reading the packed element information (history data) after a restart in
IO::DiscretizationReader::ReadHistoryData, we invoked a call to Redistribute(...)
in order to delete unnecessary element data objects which were created during UnPack.
Therefore, we used to reestablish the element row-/ and column maps in accordance
with the standard procedure for our standard distribution (see BuildElementRowColumn).
However, in problem algorithms which rely on a certain column distribution of
elements, the redistribution process is supposed to happen between Init(...)
and Setup() of the algorithm. Only then we call ReadRestart(...), because at
this point the objects relying on the parallel distribution like vectors and
matrices have to be constructed (happens during Setup).
So what happened was that after ReadRestart() the parallel distribution was reset.

To avoid this issue with this commit:

o a second version of DRT::Discretization::Redistribute is introduced, which recieves not
  only the desired node row-/ and column maps but also the desired element row-/ and column
  maps.

o the new Redistribute method is called in IO::DiscretizationReader::ReadHistoryData.

Note that on the one hand we now save some computational cost by avoiding the call to
BuildElementRowColumn; but on the other hand we temporarily increase the memory consumption
because not only the node maps have to be stored temporarily but also the element maps in
IO::DiscretizationReader::ReadHistoryData.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23381 060e22d8-9d1c-0410-868a-e3a950d63d95