Adjust indentation in a couple of files.
The new gcc-7 compiler includes a warning regarding misleading indentation which is very useful. It complains e.g. on the following code

  if (a != 17)
    a = 17;
    b = 22;

Written like this, one could think that the last statement setting b to 22 is only executed in case the 'if' evaluates to true, but this is not the case since the two statements are within {}. Thus, the compiler wants us to write the code like this:

  if (a != 17)
    a = 17;
  b = 22;

Furthermore, I avoided some warnings about ambiguous int/bool reinterpretation of the kind
  unsigned int xi, xj;
  ...
  return (xi * xj) ? xi+xj+1 : xi+xj;

  ->
  
  return (xi * xj > 0) ? xi+xj+1 : xi+xj;


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23395 060e22d8-9d1c-0410-868a-e3a950d63d95