Fix more compiler warnings with gcc-7.1.1 of the new CentOS 7.4.
These include misleading indentations, boolean/integer comparisons and a few bad habits of the kind

    if (not(whichtau_ == INPAR::FLUID::tau_taylor_hughes_zarins or
                         INPAR::FLUID::tau_taylor_hughes_zarins_wo_dt or
                         INPAR::FLUID::tau_taylor_hughes_zarins_whiting_jansen or
                         INPAR::FLUID::tau_taylor_hughes_zarins_whiting_jansen_wo_dt or
    ...

where we of course wanted to say

    if (not(whichtau_ == INPAR::FLUID::tau_taylor_hughes_zarins or
            whichtau_ == INPAR::FLUID::tau_taylor_hughes_zarins_wo_dt or
            whichtau_ == INPAR::FLUID::tau_taylor_hughes_zarins_whiting_jansen or
            whichtau_ == INPAR::FLUID::tau_taylor_hughes_zarins_whiting_jansen_wo_dt or
    ...

because the former means that we only check for the equality of the first type and then check if the other types are simply non-zero (which they are since they are enums with values larger than zero, which is why the check NEVER returned true) -> detected by: "enum constant in boolean context [-Wint-in-bool-context]".

Now there is only a severe fluid bug (stabilization parameter changed due to enum constant/boolean context) and false warnings about unitialized variables left for which I have not yet decided how to fix them.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23397 060e22d8-9d1c-0410-868a-e3a950d63d95