Clean up after r23398
The old fluid code indeed relied on the wrong behavior of whichtau_ as documented by the failing tests. We have had a broken detection mechanism of fluid stabilization parameters for many years! Thanks to the new compiler, this has finally be fixed. Furthermore, I could verify that the bug in fluid_ele_calc.cpp due to int-in-bool-context is in fact uncritical, so I fix that warning as well.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23399 060e22d8-9d1c-0410-868a-e3a950d63d95