Coupled Inverse Analysis ========================
The aim of this coupled inverse analysis is to identify the set of constitutive parameters x, 
which minimize the objective function of two experiments, 
here the tensile tests (up to now: normal inverse analysis) and the volume-pressure-change experiment.

Both data sets are normalized with respect to the data range and data size, to use both data sets equally in the identification process.
Consequently, we are able to identify the parameters, which describe both data sets best.

The jacobian and the residuum are adapted, such that the information of both experiments is included in the inverse analysis.

R = (
u_1^{mod}(x) - u^{exp}_1  \\
\vdots\\
u_{m,n}^{mod}(x) - u^{exp}_{m,n}  \\
p_1^{mod}(x) - p^{exp}_1  \\
\vdots\\
p_{n_s}^{mod}(x) - p^{exp}_{n_s}  \\
)

m,n... number of data points of tensile test
n_s ... number of data points of volume-pressure-change experiment

Compatible with Nested Parallelity (Computation of pressure values of volume-pressure-change experiment is on proc 0).

Introduction of testcase, to test this functionality.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23458 060e22d8-9d1c-0410-868a-e3a950d63d95