Constraint Optimatation =======================
A constraint is implemented to enforce, that just physiological pressure-volume-change relations are identified in the coupled inverse analysis.

With increasing pressure the volume has to increase
dp/dJ > 0

If the constraint is not fulfilled, the optimization step is not accepted and the regularization parameter μ = 2μ is adopted.
The last optimization step is repeated with this adopted regularization parameter.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23459 060e22d8-9d1c-0410-868a-e3a950d63d95