===================== Happy New Year 2018!! =====================
Model of Fractional Viscoelasticity (FSLS)
==========================================

This material offers a viscous and hyperelastic part. The model consists
of one spring in parallel to one sequential branch of a spring and a springpot.

A springpot models between a spring and a dashpot. The parameter alpha regulates 
how much damping is introduced. 
Alpha=0, means the springpot is a spring
Alpha=1, means the springpot is a dashpot; this is equal to the GenMax Material
Consequently, alpha has to be chosen between 0 and 1.

To calculate this fractional derivative the history of all previous timesteps is saved in each gauss-point.

This material is part of the viscoelasthyper-material-toolbox.

Implemented is the version prosposed in [1].
Additionally, beta is introduced, which weights the two branches (similar to beta in GenMax).
Also, no one-step theta scheme as in Enelund is necessary.
 
The input-line should read:
MAT 1 VISCO_Fract TAU 0.1 ALPHA 0.5 BETA 1

Included in viscoelasthyper-testcase.

References:
[1] Adolfson and Enelund (2003): Fractional Derivative Visocelasticity at Large Deformations
[2] Term paper of my student Liwei Luo


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23469 060e22d8-9d1c-0410-868a-e3a950d63d95