Fractional Viscoelasticity (FSLS) in Inverse Analysis =====================================================
Included fractional viscoelasticty in inverse analysis.

Constraint optimization:
The parameter alpha is not allowed to get bigger then 1.
This is checked after each optimization step.
If the constraint is not fulfilled, the optimization step is not accepted and the regularization parameter μ = 2μ is adopted.
The last optimization step is repeated with this adopted regularization parameter.

End of my personal BACI-Cleaning-Weeks :-)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23474 060e22d8-9d1c-0410-868a-e3a950d63d95