+++ data_collector +++ Added a new template version of a round robin loop which can be easily called by the non-member function DRT::CollectData. It takes the current communicator, the processor local data and collects the data over all processors, which can be found in the collected_data object. Hopefully this makes your life a little bit easier.
+++ StringConditionComponent +++
Fixed a bug in this routine for optional separator tags. When a separator is optional,
this function adds it to the condition line, if it can't be found. The bug caused that
the reader did not actually read the optional parameter, instead the next reader, expecting
the next component, received the optional tag and failed. This bug should be fixed now.

+++ drt_utils_matrix_vector_enums +++
Removed unnecassary breaks.

+++ drt_utils_parallel +++
Added a new utils function called RedistributeInAccordanceWithReference. It can be used to
adapt the parallel distribution to a potentially larger reference map, which contains the
given map as subset. In this way the distribution of the reference map and the subset map
coincide afterwards.

+++ drt_inpar +++
- Added a bunch of new input parameters. A more detailed explanation will follow.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23498 060e22d8-9d1c-0410-868a-e3a950d63d95