Fix bug in Pack/Unpack wrt. variable-sized vectors
Do not not write the `searchelements_` for now. They give us trouble
since they are a `std::vector<int>` with variable length. It looks like
they are never used anyways.

Possible remedy: When calling Unpack and stuffing data into
`searchelements_`, we first could resize it to the proper length.
In order to do so, we also have to
`AddToPack(data,static_cast<int>(searchelements_.size()))`,
so that we know the vector lenght later during Unpack().

Note: The size in Pack/Unpack needs to be given in byte,
so it should rather be
`static_cast<int>(searchelements_.size())*sizeof(int)`.

Part of #135.