Wear tests: use static redistribution (#58)
Dynamic redistribution is suspected to behave non-deterministically.
This tests passes with dynamic redistribution when executed manually,
but fails during pipeline runs on our tests machines.

For now, @aseitz proposed to switch the input flag 'PARALLEL_REDIST'
to 'static' to make the pipeline pass again. However, the dynamic case
has still to be investigated and fixed.