Enabled Contact in new structural time integration
this includes
- Frictional contact
- Penalty contact
- Condensed Lagrange multipliers
- Second order FE with basis transformations
- Penalty mesh tying
- Lagrange multiplier mesh tying (saddle point and condensed form)
- combined mesh tying and contact problems

Condensation is performed in RunPreApplyJacobianInverse(...) and
recovery in RunPostApplyJacobianInverse(...), i.e. right before and
after
we solve the linear system of equations. That way, it is ensured that
all
other involved models have already added their contributions to the
stiffness
matrix. Multiple condensations e.g. of contact and mesh tying are
handled
automatically as long as the respective interfaces do not intersect.
Intersection of interfaces is not allowed from a theoretical point of
view
since we'd need a cross-point modification to avoid being
over-constrained at
the intersection.

All functionality in the 'old' framework remain active

Selective test cases have been moved to new structural time integration

Thx to Michael Hiermeier for helpful discussions.