Fast CONTACT::AUG::TimeMonitor (#67)
Time monitor based on enum classes as template parameter. The idea is
to provide a simple and easily extendable time monitor class which can
be used to track the evaluation times of each contact interface element
without generating too much overhead.

This class is already well-tested in my local build and will be applied
in a later commit.