Introduce new core particle framework (#78)
Implementation of a new core particle framework allowing for parallel
high performance particle simulations consisting of the following
components:

- particle container managing the storage of particles including
  automatic size adaption
- particle engine managing the parallel distribution of particles of
  different types on all processors (based on a spatial decomposition
  via a recursive coordinate bisection algorithm using a binning
  strategy) and the ghosting of particles in boundary bins close to
  neighboring processors
- particle algorithm controlling the particle simulations
- particle time integration schemes: 'semi implicit euler' and 'velocity
  verlet' (both explicit)
- input of particles from dat-files (via a special input reader for
  particles) or with an input generator (core framework of input
  generator added)
- output of particles is written at runtime
- restart of particle simulations fully functional
- testing framework for particles simulations

Test cases for core particle framework functionality and unittests for
particle container classes are included.

Thanks to Jonas Eichinger, Maximilian Grill and Martin Kronbichler for
discussion, help and code being reused. Huge thanks to Georg Hammerl for
his effort in the 'old' particle framework that resulted in methodical
concepts and ideas that are partly still applied in this new particle
framework!