Add SPH interaction to particle framework (#81)
Add a smoothed particle hydrodynamics (SPH) interaction framework to the
new particle framework.

Note that not exactly the same functionality of the old framework is
incorporated in the new framework. The old SPH implementation is very
experimental. @chrmeier and @sfuchs discussed about the features and
method to be included in the new framework.

Unittests and test cases covering the new functionality are included.