Move functionality to binning strategy
As mentioned in merge request !60 moved a function to determine the
centers of all bins from particle engine to binning strategy.

In addition two methods to determine the minimum resp. the maximum bin
size with respect to a coordinate direction are added.

Discussed with @jeichinger in order to provide the functionality of
distributing bins via recursive coordinate bisection to other users.