No update of active set in predictor step (#57)
The predictor does not change the interface displacements, at least
in most cases. Hence, we can suppress the update of the active set
in the predictor step.

This is particularly important to pass the patch test, where the
entire slave interface is initialized as 'Active' and the gap is
exactly zero. An update of the active set would spoil these initial
conditions by setting slave nodes to inactive. Then, rigid body modes
would come into play and make the patchtest fail.