Issue with the Nitsche strategy
I removed the Evaluate method from the Nitsche strategy. It has never
been my intention that there are more than one Evaluate method, since
multiple versions ask for multiple points which have to be adapted if a
new action is added. Therefore, I resolved this issue and removed the
virtual tag in the abstract strategy.