Set-up framework for contact constitutivelaws #125
This commit introduces the possibility to relate
the contact gap to the contact pressure by a
constitutive law. This law should depend on micro
data like the roughness data. For now only
(nonlinear) functions are implemented to describe
the relation, but in theory it is also possible
to include full a Boundary Element Method on the
micro scale to compute the gap-pressure relation.