Modify particle transfer safety check (#133)
The particle positions after the last particle transfer are now stored
as a particle state in the respective particle container of each phase.
This way, the correct position values are directly linked to the
particles and are in the case of phase changes also transfered to the
new particle type container.