Add random noise to init particle position (#137)
Implement a method in particle time integration class, that adds a
random noise to the initial particle positions. Only particle types that
are subject to time integration are processed. The added random noise
may not be larger than the minimum relevant bin size.

In the particle algorithm, since the particle positions change, the
connectivity needs to be checked and in case of larger displacements
renewed.

Add a missing break statement in particle_enums.cpp file.