Introduce a pure virtual interface to wall handler
Introduce a pure virtual interface to provide access to the particle
wall handler. The particle algorithm class holds a shared pointer to an
particle wall base class object, thus having full access to the object
itself (in terms of public methods). An interface can be handed to other
class to provide restricted access to the particle wall handler, e.g.,
in particle interaction classes.

part of #191