Solver for contact: update re-use policy
Introduce a more careful check for changes in the active set
since we have to rebuild the preconditioner everytime the active
set changes. Now, we actually compare the current map of active
DOFs with the one from the previous application of the preconditioner.
This is the only _true_ indicator for changes in the active set.

Closes #167