Optimize data structure: particles in bins
The relation of particles to bins is stored in a std::vector, that is
indexed via the local ids of the bins with respect to the bin col map.
The particles per bin are stored in a std::vector with members std::pair
consisting of type enum and local index.

part of #115