Optimize data structure: particle neighbor pairs
In SPH and DEM the particle neighbor pairs are stored using std::vectors
instead of std::map. The particle type and index of interacting particle
i are indexed in the std::vectors via the type enum and the local
particle id.

Note: adapted results of a test case (the order of the particle pair
evaluations is different than before)

part of #115