Introduce abstract binary tree interface
An abstract interface, all implemented binary trees are inheriting from
, is introduced.
Cleared the constructors of the binary trees.
All work that was done in the constructor so far, is now moved to an
initialization method.
Removed commented out and obviously depreciated function call
SearchContactInit(..), as function is not available any more.

part of #173