Reduce (half surrounding) bin to bin relation
Adding a data structure that stores half of the surrounding 26 bins of
an owned bin (so 13 bins), the bin itself and all ghosted neighboring
bins.

The particle neighbors are build using the reduced bin neighbors, each
particle pair is checked once and added for both particles (overlapping)
to the particle neighbors.

Thus computational effort of method `BuildParticleToParticleNeighbors()`
is reduced by a factor of 2.

note: adapting result of sensitive SPH couette flow test case, due to
a different order of particle neighbor pair evaluations

part of #115