Bug fix in function evaluate in particle framework
In the evaluation of functions for initial and dirichlet boundary
conditions in the particle framework the pointer to the position where
the functions need to be evaluated was wrong. Instead of handing over a
pointer to the x value of the position of a particle for all evaluations
of function components, the pointer was shifted for each component
evaluation resulting in wrong positions being evaluated.

This bug does not affect one dimensional test cases or test cases where
the function is independent of the position. However, in test
`particle_dem_3d_normalcontact_linspringdamp-p2` until now the initial
velocity condition was evaluated resulting in wrong initial particle
velocities that do not match the given initial position and the given
functions. Besides that, the initial velocities differed if the test
case was run on a different number of processors, due to a different
distribution of particles among the processors and thus different arrays
storing the position of particles.

In order to keep changes small, the test was adapted in such a way, that
the initial velocities of the particles remain the same than before (if
executed on 2 processors), resulting in modified functions in the input
file but unchanged results.

Thanks to @mayrmt and @istein for revealing that bug. Now the test gives
the same result independently of the number of processors.

related to #179