Introduce self contact for non-smooth geometries
This commit introduces an unbiased self contact formulation for non-
smooth geometries based on a Nitsche method that does not rely on a
master-slave sorting.

Therefore I introduced:
- two_half_pass-approach: indicates that contact surface is not
separated in slave and master (and only integrated over slave side) but
instead the whole surface is "named" slave and integrated
- added a function needed for self contact on non-smooth geometries
that does checks in reference coordinates
- self contact search is changed such that a better parallelization is
possible (surface elements can only be combined to root nodes if owned
by the same processor)
- minor cleanup
- enabled self contact for nitsche contact formulation
- enabled two half pass approach for nitsche contact formulation
- added flags for the TwoHalfPass approach and the
RefConfCheckNonSmoothSelfContact
- added functionality needed for nitsche contact to tet4-elements
- added test cases to test new functionality

Moreover some member functions of the self binary tree class are
restructured, such that the unbiased self binary implementation can
inherit from it and as little code redundancy as possible has to be
introduce to add this additional functionality.

part of #155