Add unit tests for oxygen partial pressure
part of #184

Test the numerical inversion of the oxygen concentration. We know the
function to calculate the oxygen concentration from the oxygen partial
pressure (oxygen-hemoglobin dissociation curve). However, in the tumor
growth framework, we have mass fractions as primary variables and need
to get the partial pressure in terms of these mass fractions resp.
concentrations. So, the relation has to be inverted numerically. In this
commit, unit tests have been added to test the correct inversion. This
is done in the following way: We first define a partial pressure and get
the respective oxygen concentration for this pressure, then we can
invert the relation numerically with this concentration and get back a
partial pressure. If the inversion is correct, the partial pressure
should be the same (or the error should be below a reasonable limit) as
the initially defined partial pressure.