Switch to pairwise particle interaction evaluation
In both SPH and DEM the particle interactions are now evaluated in a
pairwise manner, meaning the contributions of both particles to each
other are evaluated in the same step.

In DEM the forces of two interacting particles are symmetric due to
Newton's third law. In SPH the interaction contributions of two
particles are in general non-symmetric, however some terms then need to
be evaluated just once, so that the overall computational effort can be
reduced.

In order to enable pairwise evaluation of interactions the data
structure storing the particle pairs in DEM and SPH named
`neighborpairdata_` are modified. The interaction routines are then
adapted to account for the different data structure.

additional changes:
- add const type qualifier where appropriate to indicate that certain
  variables are not modified
- avoid repeated division operations in particle interaction evaluation

note:
- adapting results of sensitive SPH couette flow test case, due to a
  different order of particle neighbor pair evaluations
- adapting results of free surface dambreak test cases, due to a minor
  bug that is now erased
- unit tests are adapted in accordance with the functions to be tested

part of #115